LayoutTests/inspector-protocol/nmi-webaudio-expected.txt

PASS: received size for type Audio is more than 15000000
